Cast to int for MSVC build

+++ Filters/hqx/HqxCommon.h
@@ -60,9 +60,9 @@ static inline u_int32_t rgb_to_yuv(u_int32_t c)
 
 /* Test if there is difference in color */
 static inline int yuv_diff(u_int32_t yuv1, u_int32_t yuv2) {
-    return (( abs((yuv1 & Ymask) - (yuv2 & Ymask)) > trY ) ||
-            ( abs((yuv1 & Umask) - (yuv2 & Umask)) > trU ) ||
-            ( abs((yuv1 & Vmask) - (yuv2 & Vmask)) > trV ) );
+    return (( abs( (int)(yuv1 & Ymask) - (int)(yuv2 & Ymask)) > trY ) ||
+            ( abs((int)(yuv1 & Umask) - (int)(yuv2 & Umask)) > trU ) ||
+            ( abs((int)(yuv1 & Vmask) - (int)(yuv2 & Vmask)) > trV ) );
 }
